[feat] Holy shit that workedgit add -A Using multiple class inheretence levels to share common database functions & variables up until a central manager
